feat: Add Dice and DiceHand components
The code changes introduce two new components, `Dice` and `DiceHand`, to the `dobbelen-frontend` project. These components are used to render dice with a specific number of dots. The `Dice` component takes a `value` prop representing the number of dots to render, while the `DiceHand` component takes an array of `diceValues` to render multiple dice.